add:bot, is working

+++ bot/src/main/java/edu/java/bot/model/BotUser.java
@@ -1,8 +1,8 @@
 package edu.java.bot.model;
 
 public record BotUser(
-    int id,
-    int chatId,
+    long id,
+    long chatId,
     String name,
     boolean isBot
 ) {
